Split tables.py into several files
Also, remove the test alphabetical order
