+++ pokedex/tests/test_schema.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from sqlalchemy import Column, Integer, String, create_engine
+from sqlalchemy import Column, Integer, String, Unicode, create_engine
 from sqlalchemy.orm import class_mapper, joinedload, sessionmaker
 from sqlalchemy.orm.session import Session
 from sqlalchemy.ext.declarative import declarative_base
@@ -30,16 +30,6 @@ def test_variable_names(varname):
 def test_variable_names_2(table):
     """We also want all of the tables exported"""
     assert getattr(tables, table.__name__) is table
-
-def test_class_order():
-    """The declarative classes should be defined in alphabetical order.
-    Except for Language which should be first.
-    """
-    class_names = [table.__name__ for table in tables.mapped_classes]
-    def key(name):
-        return name != 'Language', name
-    print [(a,b) for (a,b) in zip(class_names, sorted(class_names, key=key)) if a!=b]
-    assert class_names == sorted(class_names, key=key)
 
 def test_i18n_table_creation():
     """Creates and manipulates a magical i18n table, completely independent of
@@ -190,7 +180,7 @@ def test_texts(cls):
                 pytest.fail('%s: official text with bad format' % column)
             text_columns.append(column)
         else:
-            if isinstance(column.type, tables.Unicode):
+            if isinstance(column.type, Unicode):
                 pytest.fail('%s: text column without format' % column)
         if column.name == 'name' and format != 'plaintext':
             pytest.fail('%s: non-plaintext name' % column)
